Change 'eventing module' to 'events module'
Eventing is a type of horse riding event and I can't find any cases of
it being commonly used to refer to handling events in code.

Also, I added periods after text in a list because it looks to me like
that more closely matches what is done in other lists in the docs
section.